git: add function to parse message trailers
This is a pure python implementation of the message trailer parsing
algorithm in git (and libgit2). It is intended for use on finalized
commit messages only. Lines starting with comments are not ignored.

The function returns a list of pairs (name, value) where name is the
name of the trailer. Trailer values may span over multiple lines.

Link: https://git-scm.com/docs/git-interpret-trailers
Link: https://github.com/git/git/blob/master/trailer.c
Link: https://github.com/libgit2/libgit2/blob/main/src/trailer.c
Signed-off-by: Robin Jarry <robin@jarry.cc>