Refactor GQL webhook routing
The GQL webhooks are designed around the hub.sr.ht data model where each
hub mailing list actually represents a project-mailing-list-relation and a
specific webhook belongs to one such relation.

The web routing, however, does not fully reflect that model, as neither
the webhook payload nor the URL contain any information about the
project.

This commit sets up a new GQL webhook handler that makes the URL
project-specific. The old handler is kept around for now, so that the
existing webhooks continue to be delivered.

The `contrib/ensure-webhooks` script is extended so that it can be used
to both convert the existing GQL webhooks to the new routing layout as
well as use the new layout right away when used for its original purpose
(going from legacy to GQL webhooks).