projects: clean associated webhooks upon project deletion
When deleting a project, all associated objects (lists, trackers...) are
deleted *in the database* thanks to foreign key constraints.

However, some explicit clean-up is required for lists, repositories and
trackers: some webhooks are created in [lists,git,hg,todo].sr.ht when
the list, repository or tracker is associated to a project, and it's
left dangling when the project is deleted if we only rely on the
database to clean-up.

This patch adds this explicit clean-up step, and fixes a bug that made
[git,todo].unensure_*_webhooks a no-op instead of a deletion.

Fixes: https://todo.sr.ht/~sircmpwn/hub.sr.ht/90
Signed-off-by: Simon Martin <simon@nasilyan.com>