Handle deleted remote git repository on local removal
If the git.sr.ht repository was deleted but the project hub didn't hear
about it (e.g. due to a network partition that prevented the webhook
from being delivered), this would fail when attempting to remove the
webhook. We add a basic try..catch here which lets the operation proceed
normally.

Some other sorts of errors are swallowed up unconditionally here (such
as timeouts), which could lead to inconsistent states, but it's not the
end of the world -- the other end, the webhook receiver, can handle
those states just fine. We can narrow this in more once we have more
universal semantic error codes.